+++ resources/examples/manager/BasicAssetLibrary/plugin/BasicAssetLibrary/bal.py
@@ -30,10 +30,6 @@ from collections import namedtuple
 from urllib.parse import urlparse
 
 
-from openassetio import constants
-from openassetio.exceptions import InvalidEntityReference
-
-
 EntityInfo = namedtuple("EntityInfo", ("name"), defaults=("",))
 Entity = namedtuple("Entity", ("traits"), defaults=({},))
 
@@ -80,8 +76,8 @@ def parse_entity_ref(entity_ref: str) -> EntityInfo:
     """
     uri_parts = urlparse(entity_ref)
 
-    if not uri_parts.path:
-        raise InvalidEntityReference("Missing entity name in path component", entity_ref)
+    if len(uri_parts.path) <= 1:
+        raise InvalidBALReference("Missing entity name in path component")
 
     # path will start with a /
     name = uri_parts.path[1:]
@@ -100,7 +96,12 @@ def entity(entity_info: EntityInfo, library: dict) -> Entity:
     """
     Retrieves the Entity data addressed by the supplied EntityInfo
     """
-    return Entity(**library["entities"][entity_info.name]["versions"][-1])
+    entities_dict = library["entities"]
+    entity_dict = entities_dict.get(entity_info.name)
+    if entity_dict is None:
+        raise UnknownBALEntity()
+
+    return Entity(**entity_dict["versions"][-1])
 
 
 def managementPolicy(trait_set: set[str], library: dict) -> dict:
@@ -120,3 +121,11 @@ def managementPolicy(trait_set: set[str], library: dict) -> dict:
     else:
         policy = read_policies.get("default", {"openassetio.Managed": {}})
     return policy
+
+
+class UnknownBALEntity(Exception):
+    pass
+
+
+class InvalidBALReference(Exception):
+    pass