+++ resources/examples/manager/BasicAssetLibrary/tests/fixtures.py
@@ -93,7 +93,13 @@ fixtures = {
             "a_reference_to_a_readable_entity": f"bal:///{an_existing_entity_name}",
             "a_set_of_valid_traits": {"string", "number"},
             "a_reference_to_a_readonly_entity": f"bal:///{an_existing_entity_name}",
-            "the_error_string_for_a_reference_to_a_readonly_entity": "BAL entities are read-only"
+            "the_error_string_for_a_reference_to_a_readonly_entity": "BAL entities are read-only",
+            "a_reference_to_a_missing_entity": "bal:///missing_entity",
+            "the_error_string_for_a_reference_to_a_missing_entity": (
+                "Entity 'bal:///missing_entity' not found"),
+            "a_malformed_entity_reference": f"bal:///",
+            "the_error_string_for_a_malformed_entity_reference": (
+                f"Missing entity name in path component")
         }
     }
 }