+++ src/openassetio-core-c/private/managerApi/CManagerInterfaceAdapter.cpp
@@ -98,6 +98,16 @@ bool CManagerInterfaceAdapter::isEntityReferenceString(
   throw std::runtime_error{"Not implemented"};
 }
 
+void CManagerInterfaceAdapter::resolve(
+    [[maybe_unused]] const EntityReferences& entityReferences,
+    [[maybe_unused]] const trait::TraitSet& traitSet,
+    [[maybe_unused]] const ContextConstPtr& context,
+    [[maybe_unused]] const HostSessionPtr& hostSession,
+    [[maybe_unused]] const ManagerInterface::ResolveSuccessCallback& successCallback,
+    [[maybe_unused]] const ManagerInterface::ResolveErrorCallback& errorCallback) {
+  throw std::runtime_error{"Not implemented"};
+}
+
 }  // namespace managerApi
 }  // namespace OPENASSETIO_CORE_ABI_VERSION
 }  // namespace openassetio