+++ src/openassetio-core/include/openassetio/BatchElementError.hpp
@@ -42,7 +42,42 @@ class BatchElementError final {
   /// Possible classes of error.
   enum class ErrorCode {
     /// Fallback for uncommon errors.
-    kUnknown = OPENASSETIO_BatchErrorCode_kUnknown
+    kUnknown = OPENASSETIO_BatchErrorCode_kUnknown,
+    /**
+     * Error code used during @ref resolve "entity resolution" when the
+     * reference itself is valid, but it is not possible to retrieve
+     * data for the referenced @ref entity.
+     *
+     * This could be because it does not exist, or a read-only entity is
+     * being resolved for write. This code should not be used if an
+     * entity does not have a requested trait - simply do not set that
+     * trait in the resulting data.
+     *
+     * This code is also used during version finalisation and any other
+     * entity-based operations on a valid @ref entity_reference that
+     * fail for some reason.
+     */
+    kEntityResolutionError = OPENASSETIO_BatchErrorCode_kEntityResolutionError,
+
+    /**
+     * Error code used whenever an Entity-based action is performed on
+     * a mal-formed or unrecognized @ref entity_reference.
+     *
+     * Entity references are initially validated as part of constructing
+     * an @fqref{EntityReference} "EntityReference" object. However,
+     * that is a naive check intended to validate the general format of
+     * a reference string. It is still possible that an entity reference
+     * is invalid in the context of the entity being referenced.
+     *
+     * For example, assuming entity references are encoded as URIs, a
+     * `kInvalidEntityReference` could indicate that the query
+     * parameters suffixed to a reference string are not relevant to
+     * that particular reference. Such a check would likely require a
+     * query to the backend system. In contrast, the naive
+     * pre-validation used when constructing an `EntityReference` object
+     * is intended to be a simple in-process check.
+     */
+    kInvalidEntityReference = OPENASSETIO_BatchErrorCode_kInvalidEntityReference
   };
 
   /// Error code indicating the class of error.