+++ tests/openassetio-core-c/private/hostApi/ManagerTest.cpp
@@ -46,6 +46,7 @@ struct MockManagerInterface : trompeloeil::mock_interface<managerApi::ManagerInt
   IMPLEMENT_MOCK2(initialize);
   IMPLEMENT_CONST_MOCK3(managementPolicy);
   IMPLEMENT_CONST_MOCK2(isEntityReferenceString);
+  IMPLEMENT_MOCK6(resolve);
 };
 /**
  * Mock implementation of a HostInterface.